feat(gpio16): add GPIO16 support and integrate into button handling
* Added gpio16.cpp and gpio16.h for GPIO16 handling
* Updated platformio.ini to include common build flags for different environments
* Modified button.h to initialize and configure GPIO16 based on BUTTON_PIN

+++ src/led.h
@@ -21,7 +21,7 @@ void turnOffLed()
   digitalWrite(LED_PIN, LOW);
 }
 
-void blinkLed(int blinkInterval)
+void blinkLed(long unsigned int blinkInterval)
 {
   unsigned long currentTime = millis();
